+++ spec/dm/matchers/have_many_and_belong_to_spec.rb
@@ -8,15 +8,15 @@ describe DataMapper::Matchers::HaveManyAndBelongTo do
     end
 
     it 'fails if association does not exist' do
-      lambda { Book.should have_many_and_belong_to :authors}.should
-	fail_with "expected to have many and belong to authors"
+      lambda { Book.should have_many_and_belong_to :authors}.
+        should fail_with "expected to have many and belong to authors"
     end
   end
 
   context '#should_not' do
     it 'fails if association exists' do
-      lambda { Author.should_not have_many_and_belong_to :books}.should
-	fail_with "expected to not have many and belong to books"
+      lambda { Genre.should_not have_many_and_belong_to :books}.
+        should fail_with "expected to not have many and belong to books"
     end
   end
 end